Commit for CSS of Table of Contents
Tentative CSS for the Table of Contents, where I have a tabbed
navigation bar (IN PURE CSS!!!!!!). Some numbers probably have to be
tweaked and it hasn't been implemented yet, so bugfixes to come!

Inspiration taken from the tutorial:
http://codepen.io/wallaceerick/pen/ojtal?editors=110

I won't paste the list of things todo because I sorta forgot to
copy-paste it and this is just like, a minor commit so it's not exactly
necessary (although the former reason is probably the more important
one).