Tabs are FINISHED! Table of Contents just needs actual... content. Heh.
Ok, so the tabs are all done, and the Table of Contents just needs
filling in. I wrote a metric ton of CSS and HTML to make this work, plus
some borrowed inspiration. Font and colors are all patched up and the
<h1> element is hacked together on it's current position in a strange
mix of transform: translatey and margin-left.

TODO LIST!
[x] Finish home page
[x] Finish ToC
[ ] Make framework for individual pages
[ ] Make framework for "About"
[ ] Actually fill in content
[x] Get this list copypasted
[ ] Fill "About" page with content
[ ] Utterly shameless self-promotion